Renaming TableFields into Fields. Moving Table and Table.css into components/Table directory. Modified CSS.
