+++ src/components/Field.js
@@ -1,0 +1,22 @@
+import React from 'react';
+
+class TableFields extends React.Component {
+    getStyle = (piece) => {
+        return {
+            color: piece === piece.toLowerCase()? 'black':'white'
+        }
+    }
+
+    render() {
+        const draw = this.props.fields.reverse().map((chessField) => (
+            <div className="flexChild" style={this.getStyle(chessField.piece)} key={chessField.id}>{this.props.piece[chessField.piece]}</div>));
+
+        return (
+            <div className="flexContainer">
+                {draw}
+            </div>
+        );
+    }
+}
+
+export default TableFields;
